requirements.txt
streamlit==1.12.0
matplotlib==3.5.2
Pillow==9.0.1
torch
torchvision